[WASimUI] Data Requests table: Move default position of Index column to follow var. name; Make all columns resizeable; Save header view state between uses; Fix display of char-sized data values (were interpreted as ASCII codes); DeltaEpsilong format now only hows minimum number of digits; Add tooltips for all values in the table; Add `RequestRecord::properties` for future meta data.

+++ src/WASimUI/Utils.h
@@ -174,11 +174,11 @@ class Utils
 					break;
 				case QMetaType::Char:
 				case QMetaType::SChar:
-					v.setValue(qint8(res));
+					v.setValue(qint16((int8_t)res));  // upcast for printing
 					break;
 				case QMetaType::UChar:
 				case QMetaType::Bool:
-					v.setValue(quint8(res));
+					v.setValue(quint16((uint8_t)res));  // upcast for printing
 					break;
 				case QMetaType::Short:
 					v.setValue(qint16(res));
